+++ clinic-grpc-service/src/main/java/ru/vyrostkov/grpc/repositories/PetTypeRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.vyrostkov.grpc.models.PetType;
 @Repository
-public interface PetTypeRepository extends JpaRepository<PetType, Long> {
+public interface PetTypeRepository extends JpaRepository<PetType, Integer> {
 
     PetType findByName(String name);
 }
